# Generate date ranges in a way that is more friendly to the cache.