This commit uses rounding in order to improve the cacheability of date ranges.
For instance, `[now-1M TO now]` would actually be parsed as
`[now-1M TO now-1M/H] OR [now-1M/H TO now/H] OR [now/H TO now]`. The clause in
the middle of the disjunction has rounded bounds, which makes it more likely to
be cached. The two outer clauses are unlikely to get cached, but since the match
less than one hour of data each, they should execute quickly anyway.

In the case that the range query is already rounded at a granularity greater
than or equal to one hour, the query is left as-is. So for instance, this query:
`[now-2M/d TO now/d]` would be executed as-is without being split in 3
components like the previous query.

Closes #20106