@jpountz 2016-08-26
We discussed about the fact that it might slow down random ranges in FixitFriday. The plan is to first make sure the slow down is contained, and if this is the case merge the change as-is and work on a fix if we later get evidence that there are actual use-cases that this change hurts.